feat: 日本語コミュニケーション強制化の修正とmainブランチのマージ
- `scripts/language_check.py` にファイル名を変更し、git管理対象に含まれるように修正しました。
- main ブランチ（origin/main）の最新の変更をマージし、PNGファイルの競合を解決しました。
- AGENTS.md と .pre-commit-config.yaml の参照先を更新しました。
