+++ docs/README_DEV.md
@@ -310,6 +310,16 @@ sequenceDiagram
 
 ## 4. QL-Animation Studio
 
+### モジュール構成 (Studio)
+
+Studio は、アニメーションモジュールの開発・検証を行うためのツールです。機能ごとに以下のモジュールに分割されています。
+
+- **js/studio.js (メイン):** 全体の初期化、共通状態（`state`）の管理、各モジュールのオーケストレーション。サンプルのロードやテストの開始・停止制御。
+- **js/editor.js (エディタ):** コードエディタ（Variables, setup, draw, interaction）の制御。シンタックスハイライト、行番号（ガター）、オートインデント、検索・置換機能。
+- **js/preview.js (プレビュー):** プレビューエンジンの制御。再生/停止/一時停止/スクラブ（早送り・巻き戻し）操作、カラープリセット、速度制御、除外エリアシミュレーターのドラッグ＆ドロップ。
+- **js/metrics.js (メトリクス):** 描画パフォーマンス（Latency, Density, Change Rate）の計測とメーター表示。コンソールログの収集。
+- **js/io.js (入出力):** アニメーションコードのビルド、JSファイルとしてのダウンロード、既存ファイルのアップロード、PR作成ガイド。
+
 ### アーキテクチャ図 (Studio)
 
 ```mermaid
@@ -462,6 +472,15 @@ graph TD
 ---
 
 ## 5. QL-Category Editor
+
+### モジュール構成 (Category Editor)
+
+Category Editor は、アプリで使用するカテゴリ設定を視覚的に編集するためのツールです。
+
+- **js/category-editor.js (メイン):** 初期化、言語・テーマ設定、共通状態の管理、アニメーションエンジンのセットアップ。
+- **js/history.js (履歴):** 履歴スタック（Undo/Redo）の管理。テキスト入力の確定タイミング（blur）に合わせた履歴記録。
+- **js/ui.js (UI制御):** カテゴリリストの描画、詳細パネルのレンダリング、ドラッグ＆ドロップによる並べ替え、カラーパレット、タグ管理。
+- **js/data-io.js (データ入出力):** NDJSON形式によるインポート/エクスポート、クリップボード操作、スキーマバリデーションとの連携。
 
 ### アーキテクチャ図 (Category Editor)
